### 2022-03-05 #4225 @editorialbot
[PRE REVIEW]: BioDeg: A finite element software for the simulation of the corrosion and biodegradation process in metallic biomaterials
**Submitting author:** @mbarzegary (<a href="http://orcid.org/0000-0002-1456-0610">Mojtaba Barzegari</a>)
**Repository:** https://github.com/mbarzegary/BioDeg-UI/
**Branch with paper.md** (empty if default branch):
**Version:** v0.8
