### 2022-03-07 #4225 @mbarzegary
I think the software report is not accurate enough since `cloc` is not able to detect FreeFEM codes, which are the core of the submitted software. running `cloc`  with `--force-lang="C++",idp --force-lang="C++",edp` options to force it detect FreeFEM codes as C++ results in

```
$ cloc BioDeg-UI-main --force-lang="C++",idp --force-lang="C++",edp
      78 text files.
      74 unique files.
      16 files ignored.

Language                     files          blank        comment           code
Qt                               3              0              0           3613
C++                             27            379            180           1733
TeX                             12            602             56           1214
Markdown                         4            106              0            242
Dockerfile                       3             50              6            202
C/C++ Header                     6             30              0            136
Python                           2             36             44            116
CMake                            1             14              2             50
Bourne Shell                     2              0              1             10
YAML                             1              0              0              9
DOS Batch                        1              0              0              1
SUM:                            62           1217            289           7326
```
