### 2022-03-09 0810c7d
remove latex commands from readme file
