### 2022-03-15 #4225 @danasolav
@yangbai90, @schruste, @tjhei, @chennachaos, @kerstinvg, @thelfer, @capitalaslash, @pratikvn, @pratikvn, @jamiebramwell, @Balaje, would you be interested in reviewing this work about finite element analysis of biodegradation for [the Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

The [JOSS review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#reviewing-for-joss) focuses on evaluation of [the software](https://github.com/mbarzegary/BioDeg-UI/) as well as a [short paper](https://github.com/openjournals/joss-papers/blob/joss.04225/joss.04225/10.21105.joss.04225.pdf). JOSS reviews are usually smooth and streamlined and take place on GitHub. Here is an example of a [completed review.](https://github.com/openjournals/joss-reviews/issues/1441). Note we can be flexible regarding time needed for review.
