### 2022-03-15 8f44a93
update easy installation section in readme
