### 2022-03-23 9ea58d8
update easy installation in readme
