### 2022-03-24 b0d5687
clarify mass loss plot title to address mbarzegary/BioDeg#1
