### 2022-03-24 a7fe463
Squashed 'doc/' changes from 9cbbec8..60d514e
60d514e update easy installation section
786f81a update postprocessing section

git-subtree-dir: doc
git-subtree-split: 60d514e6162eab06fdb122d311539d8dc0e6b861
