### 2022-04-08 mbarzegary/BioDeg-UI#1 @kerstinvg
Paper/Label
Hello @mbarzegary,

I have a few suggestions regarding the Paper/Software:

- For me, the term 'scaffold' can lead to misunderstandings, since a scaffold already describes a certain component in medical application. I would suggest changing it to "Degradable Part" or something more general
- On page 4 line 108-109 in the paper probably something went wrong with the citation?
- No statement was made that the code right now is not able to simulated localised corrosion.
- Further, it was not mentioned that there are other commonly used degradation codes (e.g. phenomenological surface-based corrosion model by Grogan et al. 2011)

Thanks
