### 2022-04-08 #4281 @kerstinvg
Raised another issue regarding few things https://github.com/mbarzegary/BioDeg-UI/issues/1
