+++ paper/paper.bib
@@ -9,15 +9,15 @@
   doi = {https://doi.org/10.1016/j.corsci.2021.109674},
 }
 
-@article{barzegari2021hpc,
-      author = {Mojtaba Barzegari and Liesbet Geris},
-      title ={Highly scalable numerical simulation of coupled reaction–diffusion systems with moving interfaces},
-      journal = {The International Journal of High Performance Computing Applications},
-      volume = {0},
-      number = {0},
-      pages = {10943420211045939},
-      year = {0},
-      doi = {10.1177/10943420211045939},
+@article{barzegari2022,
+  author = {Mojtaba Barzegari and Liesbet Geris},
+  title ={Highly scalable numerical simulation of coupled reaction–diffusion systems with moving interfaces},
+  journal = {The International Journal of High Performance Computing Applications},
+  volume = {36},
+  number = {2},
+  pages = {198-213},
+  year = {2022},
+  doi = {10.1177/10943420211045939},
 }
 
 @Article{Zheng2014,
