+++ paper/paper.bib
@@ -6,7 +6,7 @@
   pages = {109674},
   year = {2021},
   issn = {0010-938X},
-  doi = {https://doi.org/10.1016/j.corsci.2021.109674},
+  doi = {10.1016/j.corsci.2021.109674},
 }
 
 @article{barzegari2022,
@@ -108,7 +108,8 @@
   journal={The visualization handbook},
   volume={717},
   year={2005},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/b978-012387582-2/50038-1},
 }
 
 @article{Dapogny2014,
@@ -240,9 +241,8 @@ volume = {72},
 number = {4},
 pages = {610-619},
 keywords = {anodic dissolution, electrochemical machining, electrochemical simulations, high current densities, steric effects},
-doi = {https://doi.org/10.1002/maco.202012051},
+doi = {10.1002/maco.202012051},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/maco.202012051},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/maco.202012051},
-abstract = {Abstract Electrochemical machining (ECM) and electropolishing are examples of anodic dissolution used in a wide variety of applications. Proper simulations of ECM require the establishment of a so-called gel layer of liquid metal salt, responsible for the depletion of water close to the surface such that the oxygen evolution is reduced. However, this effect is neglected in many simulation efforts, although it is significant to improve the accuracy of the machining process. Some authors consider this effect by an engineered “water-repelling” function that depends on the local metal ion concentration. As an alternative, we discuss the theoretical description and the implementation of modified Poisson–Nernst–Planck equations taking into account expressions in which each ion has its own maximum concentration and corresponding steric limit. We investigate a basic system of water, sodium chloride, and iron ions. Simulations pending on the local metal ion concentration for current densities show that the modified model works; steric effects become important for large current densities and the formation of a resistive gel layer is responsible for the reduction of water at the electrode. We observe that the electrostatic potential resulting from space charge effects is not affected by implementing the different species sizes. Further research is needed to explore the applicability of industrial simulation tools.},
 year = {2021}
 }
