### 2022-04-08 55a2065
change scaffold term in the paper and software ui #1
