+++ paper/paper.md
@@ -24,13 +24,13 @@ bibliography: paper.bib
 
 # Summary
 
-BioDeg is an open source software written in FreeFEM (a domain-specific language for finite element programming), C++, and Python for modeling the degradation of metallic biomaterials and simulating the biodegradation behavior of medical devices, implants, and scaffolds in corrosion experiments. The underlying mathematical and computational models are already validated in the previous contributions [@barzegari2021; @barzegari2022] by comparing the predictions made by the code with the experimentally obtained quantities. BioDeg supports simulating the change of morphology of the material block/implant/scaffold, release of materials, formation of surface corrosion products, the effect of the surrounding environment (such as various electrolyte solutions and chemical components as well as the presence of fluid flow), and change of other quantities (such as pH). It features command-line and graphical interfaces, being available on all major operating systems and platforms. BioDeg is designed with open standards in mind and as a result, can be easily integrated into other established workflows such as topology optimization and tissue growth models.
+BioDeg is an open source software written in FreeFEM (a domain-specific language for finite element programming), C++, and Python for modeling the degradation of metallic biomaterials and simulating the biodegradation behavior of medical devices and implants in corrosion experiments. The underlying mathematical and computational models are already validated in the previous contributions [@barzegari2021; @barzegari2022] by comparing the predictions made by the code with the experimentally obtained quantities. BioDeg supports simulating the change of morphology of the biodegradable part, release of materials, formation of surface corrosion products, the effect of the surrounding environment (such as various electrolyte solutions and chemical components as well as the presence of fluid flow), and change of other quantities (such as pH). It features command-line and graphical interfaces, being available on all major operating systems and platforms. BioDeg is designed with open standards in mind and as a result, can be easily integrated into other established workflows such as topology optimization and tissue growth models.
 
 # Statement of need
 
 Biomaterials, the substances we put inside the body to replace/repair a lost functionality, can be classified into two categories from a biodegradation perspective: bio-inert and biodegradable. While the former type has a few interactions with its surrounding environment, the latter gradually disappears and gets absorbed by the body [@Zheng2014; @Chen2014]. Although bio-inert biomaterials show a great performance especially in fixation applications, they bring an important problem into play: they remain in the body forever or require additional surgery to remove them. Biodegradable materials do not have this problem, and in the case of metallic biomaterials, they also provide a suitable mechanical stability profile. However, taking advantage of them requires tuning the degradation parameters and material release rate [@Zhao2017].
 
-This problem is usually investigated by conducting *in-vitro* (in the lab) and *in-vivo* (in a living system) tests of biodegradable metallic scaffolds and implants, which requires conducting multiple experiments for different scenarios and situations and spending a lot of resources. Developing a quantitative mathematical model of the degradation process is a proper solution to this issue by allowing researchers to study the biodegradation behavior of any desired implant *in-silico* (in the computer) prior to conducting any *in-vitro* or *in-vivo* experiments. Developed mathematical models can be simulated using efficient numerical methods such as the finite element method.
+This problem is usually investigated by conducting *in-vitro* (in the lab) and *in-vivo* (in a living system) tests of biodegradable metallic implants, which requires conducting multiple experiments for different scenarios and situations and spending a lot of resources. Developing a quantitative mathematical model of the degradation process is a proper solution to this issue by allowing researchers to study the biodegradation behavior of any desired implant *in-silico* (in the computer) prior to conducting any *in-vitro* or *in-vivo* experiments. Developed mathematical models can be simulated using efficient numerical methods such as the finite element method.
 
[…]
