### 2022-04-11 dbc846e
Squashed 'doc/' changes from 60d514e..e67b827
e67b827 change scaffold term to biodegradable part for mbarzegary/BioDeg-UI#1

git-subtree-dir: doc
git-subtree-split: e67b827605d5c6013cd05799ded96746b6192c75
