+++ manual/intro.tex
@@ -1,9 +1,9 @@
-\biodeg{} is an open-source software written in FreeFEM (a domain-specific language for finite element programming), C++, and Python for modeling the degradation of metallic biomaterials and simulating the biodegradation behavior of medical devices, implants, and scaffolds in corrosion experiments. It can handle any geometry of desire and supports parallel computing to simulate large-scale models.
+\biodeg{} is an open-source software written in FreeFEM (a domain-specific language for finite element programming), C++, and Python for modeling the degradation of metallic biomaterials and simulating the biodegradation behavior of medical devices and implants in corrosion experiments. It can handle any geometry of desire and supports parallel computing to simulate large-scale models.
 
 \subsection{Authors}
 \label{sec:authors}
-\biodeg{} is developed by the \href{http://www.biomech.ulg.ac.be/}{Biomechanics Research group at KU Leuven and University of Liege}. The code is currently maintained by its principal developer, 
-who manages the development of the mathematical models and the core functionalities. 
+\biodeg{} is developed by the \href{http://www.biomech.ulg.ac.be/}{Biomechanics Research group at KU Leuven and University of Liege}. The code is currently maintained by its principal developer,
+who manages the development of the mathematical models and the core functionalities.
 
 \paragraph*{Principal developer}
 \begin{itemize}
@@ -34,5 +34,5 @@ The development of \biodeg{} open-source code is financially supported by the Pr
 
 \subsection{Referencing \biodeg{}}
 
-Please refer to \href{https://github.com/mbarzegary/BioDeg}{\biodeg{} repository}, section "Publications and referencing" to properly cite the use of 
-\biodeg{} in your scientific work. 
+Please refer to \href{https://github.com/mbarzegary/BioDeg}{\biodeg{} repository}, section "Publications and referencing" to properly cite the use of
+\biodeg{} in your scientific work.
