### 2022-04-11 mbarzegary/BioDeg-UI#1 @mbarzegary
hi @kerstinvg,

thank you for the constructive comments. I applied a couple of fixes (mentioned above) to address your comments.

there was a fundamental mistake in the way I had the references in the paper text, one of which was pointed out by you. actually, I had already cited Grogan work and a couple of other similar codes, but due to this mistake (using comma instead of semicolon to refer to multiple references), they were not in the compiled paper. now the issue is resolved (https://github.com/mbarzegary/BioDeg-UI/commit/342e406f83fd40d14e8b5d02613b20d556e186ec), so they should appear in the final text.
