### 2022-04-14 #4281 @yangbai90
Hi @mbarzegary : from my side, everything is fine except the autotest part. I'm wondering, is there any autotest for your package? I believe the freeFem itself already contains the autotest, how about the BioDeg?
