### 2022-04-15 #4281 @mbarzegary
hi @yangbai90 thank you for asking this. yes, BioDeg has actually some tests already implemented using google test framework. you can see them under the `tests` directory. for CI purposes, I created a docker image (pushed to docker hub) with Qt and FreeFEM built inside. this docker image is used after each commit (more precisely, each git push) to build BioDeg and run the tests using the Travis CI service. this has been done using the `.travis.yml`, in which the script `buildWithDocker.sh` is called to run the build and tests. the result of this is reflected on the build badge in the top of the readme file.

I'm not very happy with Travis at this moment since it has put some limitation on the duration of CI jobs and the credits, so I am considering switching to GitHub actions for CI stuff.
