+++ .github/workflows/main.yml
@@ -1,0 +1,35 @@
+# This is a basic workflow to help you get started with Actions
+
+name: CI
+
+# Controls when the workflow will run
+on:
+  # Triggers the workflow on push or pull request events but only for the main branch
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+
+  # Allows you to run this workflow manually from the Actions tab
+  workflow_dispatch:
+
+# A workflow run is made up of one or more jobs that can run sequentially or in parallel
+jobs:
+  # This workflow contains a single job called "build"
+  build:
+    # The type of runner that the job will run on
+    runs-on: ubuntu-latest
+
+    # Steps represent a sequence of tasks that will be executed as part of the job
+    steps:
+      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
+      - uses: actions/checkout@v3
+
+      # Pulls a Docker image and builds the software inside it
+      - name: Build and test BioDeg
+        uses: addnab/docker-run-action@v3
+        with:
+          registry: docker.io
+          image: mbarzegary/qt-5.15.2-freefem-4.10:latest
+          options: -v "$GITHUB_WORKSPACE":/data 
+          run: /data/buildWithDocker.sh
