### 2022-04-15 60440a2
update github actions config file
