### 2022-04-20 19746a8
Squashed 'core/' content from commit 7a0c805
git-subtree-dir: core
git-subtree-split: 7a0c8053920a5150da5ea26ad8eaaf32f776b6f7
