### 2022-04-20 #4281 @mbarzegary
thank you so much @kerstinvg and @yangbai90 for your constructive review comments.

@danasolav I created a new tagged release and archive:
the version number is v0.8.1: https://github.com/mbarzegary/BioDeg-UI/releases/tag/v0.8.1
the archive DOI is 10.5281/zenodo.6472450: https://doi.org/10.5281/zenodo.6472450
