### 2022-05-02 mbarzegary/BioDeg-UI#2 @danasolav
Paper review
Hi @mbarzegary,

Below are a few requests/comments regarding the paper:

1. Line 49: "a limited set of features". Can you please be more specific? In what sense is it limited? What kind of features exist in BioDeg and/or MuPhys that are unavailable in COMSOL?
2. Line 51: Please add the affiliation of the SURF research group, for completeness.
3. Line 52: what does "restricted access" mean? Is it a commercial software? Is it closed-source?
4. Figure 1: Please add more details. For example: what the different sub-figures represent (which time frames are shown?), what are the dimensions of the screw? The properties of the screw and the surrounding? Is this example provided along with the code?
5. Line 102: Please provide further explanation regarding the mesh refinement, and the elements in general. What is the mesh refinement based on? How does the preprocessor determines if and how much mesh refinement is required? What types of elements are supported?
6. Line 147: It is not clear what "the rest of the postprocessing process" is and if it relies on processes executed in ParaView.
7. Line 152: Please add a reference to BioDeg-doc.

Thanks
