### 2022-05-02 #4281 @danasolav
@mbarzegary, your submission is almost ready for publication. Please refer to the my comments regarding the paper in this [issue](https://github.com/mbarzegary/BioDeg-UI/issues/2).
