### 2022-05-19 22a3563
update figure caption in the paper #2
