+++ paper/paper.md
@@ -46,7 +46,7 @@ From a general point of view, material modeling techniques are categorized as ei
 
 BioDeg works based on the principles of deterministic modeling for predicting the biodegradation and corrosion of metallic materials. Several studies have demonstrated that this type of modeling has a high potential in representing the biodegradation and corrosion phenomena [@Dolgikh2019; @Bajger2016; @SanzHerrera2018]. BioDeg is built upon a mechanistic formulation of the biodegradation process [@barzegari2021]. The developed model captures the release of metallic ions, changes in pH, the formation of a protective film, the dissolution of this film in presence of different ions, and the effect of fluid flow of the surrounding solution. This has been accomplished by deriving a system of time-dependent reaction-diffusion-convection partial differential equations (PDEs) from the underlying oxidation-reduction reactions. The level set formalism [@RonaldFedkiw2002] was employed to track the biodegradation interface between the biomaterial and its surroundings, enabling the model to monitor the morphological changes of the investigated implant/device during the biodegradation process. As an example of BioDeg output, figure \ref{screw} demonstrates the simulation results of the biodegradation of an orthopedics screw, in which the white body and the color contour show the degrading shape of the screw and concentration of metallic ions as they are released over time, respectively. The current model is able to simulate diffusion-based corrosion only, and localized corrosion is not supported yet, but further development is being done to make the model capable of simulating non-uniform corrosion as well.
 
-![Simulation results of the degradation process of a screw. The contours display the concentration of metallic ions on a cross-section view of the medium beside the moving surface of the screw.  \label{screw}](screw_degradation.png)
+![Simulation results of the degradation process of a screw made of magnesium inside a saline (NaCl) solution, leading to a high rate of degradation. The contours display the concentration of magnesium ions on a cross-section view of the medium beside the moving surface of the screw. The screw was selected to be small (1 millimeter) so that the effect of degradation can be observed in a smaller time window. Each sub-figure shows a time difference of 24 hours. This example is provided along with the code. \label{screw}](screw_degradation.png)
 
 ## Implementation
 
