### 2022-05-19 3099e13
add reference to biodeg-doc in joss paper #2
