+++ paper/paper.bib
@@ -246,3 +246,12 @@ url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/maco.202012051},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/maco.202012051},
 year = {2021}
 }
+
+@misc{BioDegDoc,
+  doi = {10.5281/ZENODO.6564024},
+  url = {https://zenodo.org/record/6564024},
+  author = {Mojtaba Barzegari and Liesbet Geris},
+  title = {BioDeg corrosion and biodegradation simulation software: user manual and theory guide},
+  publisher = {Zenodo},
+  year = {2022}
+}
