### 2022-05-19 ea2ba33
fix issues related to surf group on the joss paper #2
