+++ paper/paper.md
@@ -34,7 +34,7 @@ This problem is usually investigated by conducting *in-vitro* (in the lab) and *
 
 Although there is a wide variety of relevant theoretical studies in the field of corrosion and biodegradation simulation [@Gastaldi2011; @Wilder2014; @Grogan2014; @Bajger2016; @Dolgikh2019; @Hoche2014; @SanzHerrera2018], to the best of the authors' knowledge, none of the developed models has been publicly released together with their code to be evaluated by the scientific community. It means that none of the previous contributions are available as a software tool for relevant users such as researchers in biomaterials science, regenerative medicine (tissue engineering), and biomedical engineering.
 
-Moreover, the available tools used to implement aforementioned models are not open source or have the issue of restricted access. For example, COMSOL multiphysics (COMSOL Inc., USA) has a module for mechanistic corrosion simulation with a limited set of features. Another example in this regard is the MuPhyS software [@Dolgikh2019; @Christoph2021] developed by the SURF research group. Despite being able to perform mechanistic biodegradation simulations, MuPhys has a restricted access and is not publicly available to general users.
+Moreover, the available tools used to implement aforementioned models are not open source or have the issue of restricted access. For example, COMSOL multiphysics (COMSOL Inc., USA) has a module for mechanistic corrosion simulation with a limited set of features. Another example in this regard is the MuPhyS software [@Dolgikh2019; @Christoph2021] developed by the research group of Electrochemical and Surface Engineering (SURF), Vrije Universiteit Brussel, Brussels, Belgium. Despite being able to perform mechanistic biodegradation simulations, MuPhys is a closed-source software for in-house applications and is not publicly available to general users.
 
 BioDeg is a fully open source biodegradation and corrosion simulation software, currently validated for magnesium biomaterials [@barzegari2021]. A cross-platform and easy-to-use user interface makes it possible for anyone to use the program to simulate the biodegradation behavior of metallic implants with any shape of interest without concerning the technical aspects of the code. Technically speaking, BioDeg uses the finite element method to solve a set of partial differential equations derived mechanistically from the chemistry of biodegradation and utilizes high-performance computing (HPC) techniques to make the computation scalable to hundreds and thousands of computational nodes.
 
