### 2022-05-19 f5d2219
address specific comments on the joss paper #2
