### 2022-05-19 mbarzegary/BioDeg-UI#2 @mbarzegary
hi @danasolav

thank you for these constructive comments. I addressed them on the above commits.
