### 2022-06-14 #4281 @mbarzegary
@danasolav thank you. I'm so happy about it :)
I haven't changed the software since I created this tagged release, so we can use this one:

the version number is v0.8.1: https://github.com/mbarzegary/BioDeg-UI/releases/tag/v0.8.1
the archive DOI is 10.5281/zenodo.6472450: https://doi.org/10.5281/zenodo.6472450
